requirements.txt
screeninfo==0.8.1
numpy==1.24.2
easyocr==1.6.2
python-Levenshtein==0.20.9